fix(secure): do not try to detect protocol and set secure to true by default
Before this commit we were trying to detect the server's protocol and set the
secure flag accordingly. But multiple users reported wrong findings because your
server could be in HTTP while it's being proxyied to HTTPS to the enduser.

This case is impossible to handle well everywhere so we're back to documenting
how to set the secure flag according to your production/dev server if this makes
sense to the user.

fixes #425