update README.md to remove reference to /edge
the "getIronSession" seems to work for me when I remove the /edge from the import. Using the /edge resulted in an error that says. " Error: no native implementation of WebCrypto is available in current context at getCrypto". Just thought this might be a good upgrade to the documentation to help new users. Thank you so much for your work on this library!